CAMEL-19280: camel-coap: refactor: move Observer functionality to CoAPObserver.java
